requirements.txt
Flask
psycopg2-binary
gunicorn
requests
# debugpy